+++ components/external_server/tests/routes/test_metric.py
@@ -1,12 +1,14 @@
 """Unit tests for the metric routes."""
 
+from datetime import date, timedelta
 from unittest.mock import Mock, patch
 
 import requests
 
 from shared_data_model import DATA_MODEL
-from shared.model.report import Report
 from shared.utils.type import User
+
+from model.report import Report
 
 
 from routes import (
@@ -14,6 +16,7 @@ from routes import (
     delete_metric,
     post_metric_attribute,
     post_metric_copy,
+    post_metric_debt,
     post_metric_new,
     post_move_metric,
 )
@@ -34,9 +37,8 @@ from ..fixtures import (
 from ..base import DataModelTestCase, disable_logging
 
 
-@patch("bottle.request")
-class PostMetricAttributeTest(DataModelTestCase):
-    """Unit tests for the post metric attribute route."""
+class PostMetricAttributeTestCase(DataModelTestCase):  # skipcq: PTC-W0046
+    """Base class for unit tests for the post metric attribute routes."""
 
     def setUp(self):
         """Extend to set up the database."""
@@ -89,6 +91,11 @@ class PostMetricAttributeTest(DataModelTestCase):
         uuids = sorted(uuids or [REPORT_ID, SUBJECT_ID, METRIC_ID])
         description = f"John Doe changed the {description}."
         self.assertDictEqual(dict(uuids=uuids, email=email, description=description), report["delta"])
+
+
+@patch("bottle.request")
+class PostMetricAttributeTest(PostMetricAttributeTestCase):
+    """Unit tests for the post metric attribute route."""
 
     def test_post_metric_name(self, request):
         """Test that the metric name can be changed."""
@@ -354,6 +361,113 @@ class PostMetricAttributeTest(DataModelTestCase):
         self.assertEqual(dict(ok=True), post_metric_attribute(METRIC_ID2, "position", self.database))
         self.database.reports.insert_one.assert_not_called()
         self.assertEqual([METRIC_ID, METRIC_ID2], list(self.report["subjects"][SUBJECT_ID]["metrics"].keys()))
+
+
+@patch("bottle.request")
+class PostMetricDebtTest(PostMetricAttributeTestCase):
+    """Unit tests for the post metric debt route."""
+
+    def setUp(self) -> None:
+        """Set up test fixtures."""
+        super().setUp()
+        self.database.measurements.find_one.return_value = None
+
+    @patch("shared.model.measurement.iso_timestamp", new=Mock(return_value="2019-01-01"))
+    def test_turn_metric_technical_debt_on_with_existing_measurement(self, request):
+        """Test that accepting technical debt also sets the technical debt value."""
+        self.database.measurements.find_one.return_value = dict(
+            _id="id",
+            metric_uuid=METRIC_ID,
+            count=dict(value="100", status_start="2018-01-01"),
+            sources=[],
+        )
+        request.json = dict(accept_debt=True)
+        self.assertDictEqual(
+            dict(
+                end="2019-01-01",
+                sources=[],
+                start="2019-01-01",
+                metric_uuid=METRIC_ID,
+                count=dict(
+                    value=None,
+                    status=None,
+                    status_start="2018-01-01",
+                    target="0",
+                    near_target="10",
+                    debt_target="100",
+                    direction="<",
+                ),
+            ),
+            post_metric_debt(METRIC_ID, self.database),
+        )
+        updated_report = self.database.reports.insert_one.call_args[0][0]
+        desired_response_time = Report(self.DATA_MODEL, updated_report).desired_response_time("debt_target_met")
+        expected_date = date.today() + timedelta(days=desired_response_time)
+        self.assert_delta(
+            "accepted debt from 'False' to 'True' and the debt target from 'None' to '100' and the debt end date from "
+            f"'None' to '{expected_date.isoformat()}' of metric 'name' of subject 'Subject' in report 'Report'",
+            report=updated_report,
+        )
+        self.assertTrue(updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["accept_debt"])
+
+    def test_turn_metric_technical_debt_on_without_existing_measurement(self, request):
+        """Test that accepting technical debt also sets the technical debt value."""
+        request.json = dict(accept_debt=True)
+        self.assertDictEqual(dict(ok=True), post_metric_debt(METRIC_ID, self.database))
+        updated_report = self.database.reports.insert_one.call_args[0][0]
+        desired_response_time = Report(self.DATA_MODEL, updated_report).desired_response_time("debt_target_met")
+        expected_date = (date.today() + timedelta(days=desired_response_time)).isoformat()
+        self.assert_delta(
+            f"accepted debt from 'False' to 'True' and the debt end date from 'None' to '{expected_date}' "
+            "of metric 'name' of subject 'Subject' in report 'Report'",
+            report=updated_report,
+        )
+        self.assertTrue(updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["accept_debt"])
+        self.assertEqual(expected_date, updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["debt_end_date"])
+
+    def test_turn_metric_technical_debt_off_that_is_already_off(self, request):
+        """Test that turning debt off when it's already off does nothing."""
+        request.json = dict(accept_debt=False)
+        self.assertDictEqual(dict(ok=True), post_metric_debt(METRIC_ID, self.database))
+        self.database.reports.insert_one.assert_not_called()
+
+    def test_turn_metric_technical_debt_off_that_is_already_off_but_has_debt_target(self, request):
+        """Test that turning debt off when it's already off does reset debt target."""
+        self.report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["debt_target"] = "100"
+        request.json = dict(accept_debt=False)
+        self.assertDictEqual(dict(ok=True), post_metric_debt(METRIC_ID, self.database))
+        updated_report = self.database.reports.insert_one.call_args[0][0]
+        self.assert_delta(
+            "debt target from '100' to 'None' of metric 'name' of subject 'Subject' in report 'Report'",
+            report=updated_report,
+        )
+        self.assertFalse(updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["accept_debt"])
+        self.assertEqual(None, updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["debt_target"])
+
+    def test_turn_metric_technical_debt_off_that_is_already_off_but_has_debt_end_date(self, request):
+        """Test that turning debt off when it's already off does reset debt end date."""
+        self.report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["debt_end_date"] = "2000-01-01"
+        request.json = dict(accept_debt=False)
+        self.assertDictEqual(dict(ok=True), post_metric_debt(METRIC_ID, self.database))
+        updated_report = self.database.reports.insert_one.call_args[0][0]
+        self.assert_delta(
+            "debt end date from '2000-01-01' to 'None' of metric 'name' of subject 'Subject' in report 'Report'",
+            report=updated_report,
+        )
+        self.assertFalse(updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["accept_debt"])
+        self.assertFalse(updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["debt_end_date"])
+
+    def test_turn_metric_technical_debt_off_without_existing_measurement(self, request):
+        """Test turning technical debt off."""
+        self.report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["accept_debt"] = True
+        request.json = dict(accept_debt=False)
+        self.assertDictEqual(dict(ok=True), post_metric_debt(METRIC_ID, self.database))
+        updated_report = self.database.reports.insert_one.call_args[0][0]
+        self.assert_delta(
+            "accepted debt from 'True' to 'False' of metric 'name' of subject 'Subject' in report 'Report'",
+            report=updated_report,
+        )
+        self.assertFalse(updated_report["subjects"][SUBJECT_ID]["metrics"][METRIC_ID]["accept_debt"])
 
 
 class MetricTest(DataModelTestCase):